Added functioning search bar and changed the structure to exclude Layout.jsx. The layout moved to App.jsx.
