+++ triplejgamestoreApp/package.json
@@ -14,6 +14,7 @@
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^5.16.4",
     "@mui/material": "^5.16.4",
+    "prop-types": "^15.8.1",
     "react": "^18.3.1",
     "react-auth-kit": "^3.1.3",
     "react-dom": "^18.3.1",
