+++ triplejgamestoreApp/src/components/GameOverview.jsx
@@ -8,12 +8,12 @@ const GameOverview = () => {
     const [loading, setLoading] = useState(true)
     const { id } = useParams()
     const navigate = useNavigate()
-    const API_URL = 'placeholder URL'                   //Remember to replace with database url
+    const API_URL = 'https://triplej-gamestore-2bf9fca17274.herokuapp.com/api/games' //Games URL
 
     useEffect(() => {
         async function getGameData() {
             try {
-                const response = await fetch(`${API_URL}/${id}`)
+                const response = await fetch(`${API_URL}/:${id}`)
                 const data = await response.json()
                 if (!response.ok) {
                     throw new Error("Game not found")
