+++ triplejgamestoreApp/src/components/Router/Router.jsx
@@ -2,8 +2,10 @@ import { useState } from "react";
 import { Route, Routes } from "react-router-dom";
 import LoginForm from "../LoginForm";
 import RegisterForm from "../RegisterForm";
+import GameStore from "../GameStore";
+import PropTypes from 'prop-types'
 
-const Router = () => {
+const Router = ({ searchTerm }) => {
   const [token, setToken] = useState(null);
 
   return (
@@ -13,7 +15,7 @@ const Router = () => {
       <Route path="/account/login" element={<LoginForm setToken={setToken} />}/>
       <Route path="/account/register" element={<RegisterForm setToken={setToken} />}/>
 
-      <Route path="/store" element={<h1>Main store</h1>} />
+      <Route path="/store" element={<GameStore searchTerm={searchTerm}/>} />{/* Remember to pass searchTerm into gameStore */}
       <Route path="/store/:gameid" element={<h1>Game details</h1>} />
       <Route path="/account/library" element={<h1>Owned Games</h1>} />
       <Route path="/account/wishlist" element={<h1>Wishlist</h1>} />
@@ -25,5 +27,7 @@ const Router = () => {
     </Routes>
   );
 };
-
+Router.propTypes = {
+  searchTerm: PropTypes.string
+}
 export default Router;
